LayoutTests/platform/chromium-mac/media/video-src-set-expected.txt

Test that setting src attribute triggers load
EVENT(loadstart)
EXPECTED (relativeURL(video.currentSrc) == 'content/test.ogv') OK
END OF TEST
